feat: enhance Redis integration with connection hints and update email queue handling
