+++ .env.template
@@ -49,8 +49,10 @@ AWS_REGION=
 AWS_BUCKET=
 
 # Redis
+# Connection URL used by app/worker when dialing Redis.
 REDIS_URL=redis://redis:6379
-REDIS_PORT=6380
+# Optional host port mapping for Redis container access from outside Docker.
+REDIS_PUBLISHED_PORT=6380
 
 # Cronitor
 CRONITOR_API_KEY=
