+++ src/config/validateEnv.ts
@@ -15,7 +15,6 @@ const requiredKeys: EnvKey[] = [
   'awsSecretAccessKey',
   'awsRegion',
   'awsBucket',
-  'redisUrl',
 ]
 
 export function validateEnv(): void {
